+++ app/controllers/general_controller.rb
@@ -17,7 +17,7 @@ class GeneralController < ApplicationController
     medium_cache
     @locale = AlaveteliLocalization.locale
     successful_query = InfoRequestEvent.make_query_from_params( latest_status: ['successful'] )
-    @request_events = InfoRequest.recent_requests
+    @request_events = recent_request_events
     @track_thing = TrackThing.create_track_for_search_query(successful_query)
     @number_of_requests = Rails.cache.fetch(
       'frontpage/info_request_count', expires_in: 1.hour
@@ -195,5 +195,20 @@ class GeneralController < ApplicationController
     if feature_enabled?(:alaveteli_pro) && current_user && current_user.is_pro?
       redirect_to alaveteli_pro_dashboard_path
     end
+  end
+
+  def recent_request_events
+    ids = Rails.cache.fetch(
+      'frontpage/recent_requests', expires_in: 10.minutes, skip_nil: true
+    ) do
+      events = InfoRequest.recent_requests
+      events.empty? ? nil : events.map(&:id)
+    end || []
+
+    InfoRequestEvent.where(id: ids).
+      joins(:info_request).merge(InfoRequest.is_searchable).
+      includes(info_request: :public_body).
+      order(created_at: :desc).
+      load
   end
 end